backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
twilio==8.10.0
sendgrid==6.10.0
firebase-admin==6.3.0
python-dateutil==2.8.2
